fix type mismatch on parameters to a couple calls to max()
With x64, size_t is 64bit instead of 32. The literal was causing msvc to
error out from being unable to expand the template.

Note: there is still much to do for x64, which I do no plan to do now
libmad has inline asm. And there are a ton of new warnings when building
for x64.

+++ src/RageDisplay_D3D.cpp
@@ -800,8 +800,8 @@ public:
 	void Allocate( const vector<msMesh> &vMeshes )
 	{
 		using std::max;
-		m_vVertex.resize( max(1u, GetTotalVertices()) );
-		m_vTriangles.resize( max(1u, GetTotalTriangles()) );
+		m_vVertex.resize( max(size_t(1u), GetTotalVertices()) );
+		m_vTriangles.resize( max(size_t(1u), GetTotalTriangles()) );
 	}
 	void Change( const vector<msMesh> &vMeshes )
 	{
